slint_interpreter

Module ffi

Available on crate feature ffi only.

Macros

ModelAdaptorVTable_static

Instantiate a static ModelAdaptorVTable for a given type and implements vtable::HasStaticVTable<ModelAdaptorVTable> for it.

Structs

CallbackUserData

Wrap the user_data provided by the native code and call the drop function on Drop.

ComponentCompilerOpaque

ComponentDefinitionOpaque

Diagnostic

Diagnostic describes the aspects of either a warning or an error, along with its location and a description. Diagnostics are typically returned by slint::interpreter::ComponentCompiler::diagnostics() in a vector.

ModelAdaptorTO

Invariant, same as vtable::Inner: vtable and ptr has to be valid and ptr an instance matching the vtable

ModelAdaptorVTable

ModelNotifyOpaque64-bit

PropertyDescriptor

PropertyDescriptor is a simple structure that’s used to describe a property declared in .slint code. It is returned from in a vector from slint::interpreter::ComponentDefinition::properties().

StructIteratorOpaque

StructOpaque64-bit

Enums

DiagnosticLevel

DiagnosticLevel describes the severity of a diagnostic.

Traits

ModelAdaptor

Note: Was generated from the #[vtable] macro on ModelAdaptorVTable

Functions

slint_interpreter_component_compiler_build_from_path⚠

slint_interpreter_component_compiler_build_from_source⚠

slint_interpreter_component_compiler_destructor⚠

slint_interpreter_component_compiler_get_diagnostics⚠

slint_interpreter_component_compiler_get_include_paths⚠

slint_interpreter_component_compiler_get_style⚠

slint_interpreter_component_compiler_new⚠

slint_interpreter_component_compiler_set_include_paths⚠

slint_interpreter_component_compiler_set_style⚠

slint_interpreter_component_compiler_set_translation_domain⚠

slint_interpreter_component_definition_callbacks⚠

Returns the list of callback names of the component the component definition describes

slint_interpreter_component_definition_clone⚠

Construct a new Value in the given memory location

slint_interpreter_component_definition_destructor⚠

Destruct the component definition in that memory location

slint_interpreter_component_definition_functions⚠

Returns the list of function names of the component the component definition describes

slint_interpreter_component_definition_global_callbacks⚠

Returns a vector of the names of the callbacks of the specified publicly exported global singleton. Returns true if a global exists under the specified name; false otherwise.

slint_interpreter_component_definition_global_functions⚠

Returns a vector of the names of the functions of the specified publicly exported global singleton. Returns true if a global exists under the specified name; false otherwise.

slint_interpreter_component_definition_global_properties⚠

Returns a vector of the property descriptors of the properties of the specified publicly exported global singleton. Returns true if a global exists under the specified name; false otherwise.

slint_interpreter_component_definition_globals⚠

Returns a vector of strings with the names of all exported global singletons.

slint_interpreter_component_definition_name⚠

Return the name of the component definition

slint_interpreter_component_definition_properties⚠

Returns the list of properties of the component the component definition describes

slint_interpreter_component_instance_component_definition⚠

slint_interpreter_component_instance_create⚠

Instantiate an instance from a definition.

slint_interpreter_component_instance_get_global_property⚠

Get a global property. Returns a raw boxed value on success; nullptr otherwise.

slint_interpreter_component_instance_get_property

Get a property. Returns a null pointer if the property does not exist.

slint_interpreter_component_instance_invoke

Invoke a callback or function. Returns raw boxed value on success and null ptr on failure.

slint_interpreter_component_instance_invoke_global⚠

Invoke a global callback or function. Returns raw boxed value on success; nullptr otherwise.

slint_interpreter_component_instance_set_callback⚠

Set a handler for the callback. The callback function must initialize the ret (the ret passed to the callback is initialized and is assumed initialized after the function)

slint_interpreter_component_instance_set_global_callback⚠

The callback function must initialize the ret (the ret passed to the callback is initialized and is assumed initialized after the function)

slint_interpreter_component_instance_set_global_property

slint_interpreter_component_instance_set_property

slint_interpreter_component_instance_show

Show or hide

slint_interpreter_component_instance_window⚠

Return a window for the component

slint_interpreter_model_notify_destructor⚠

Destruct the value in that memory location

slint_interpreter_model_notify_new⚠

Construct a new ModelNotifyNotify in the given memory region

slint_interpreter_model_notify_reset⚠

slint_interpreter_model_notify_row_added⚠

slint_interpreter_model_notify_row_changed⚠

slint_interpreter_model_notify_row_removed⚠

slint_interpreter_struct_clone⚠

Construct a new Struct in the given memory location

slint_interpreter_struct_destructor⚠

Destruct the struct in that memory location

slint_interpreter_struct_get_field

slint_interpreter_struct_iterator_destructor⚠

slint_interpreter_struct_iterator_next⚠

Advance the iterator and return the next value, or a null pointer

slint_interpreter_struct_make_iter

slint_interpreter_struct_new⚠

Construct a new Struct in the given memory location

slint_interpreter_struct_set_field

slint_interpreter_value_clone

Construct a new Value in the given memory location

slint_interpreter_value_destructor

Destruct the value in that memory location

slint_interpreter_value_enum_to_string

slint_interpreter_value_eq

slint_interpreter_value_new

Construct a new Value in the given memory location

slint_interpreter_value_new_array_model

Construct a new Value in the given memory location as array model

slint_interpreter_value_new_bool

Construct a new Value in the given memory location as bool

slint_interpreter_value_new_brush

Construct a new Value in the given memory location as Brush

slint_interpreter_value_new_double

Construct a new Value in the given memory location as double

slint_interpreter_value_new_enum

slint_interpreter_value_new_image

Construct a new Value in the given memory location as image

slint_interpreter_value_new_model⚠

Construct a new Value containing a model in the given memory location

slint_interpreter_value_new_string

Construct a new Value in the given memory location as string

slint_interpreter_value_new_struct

Construct a new Value in the given memory location as Struct

slint_interpreter_value_to_array

Extracts a SharedVector<ValueOpaque> out of the given value val, writes that into the out parameter and returns true; returns false if the value does not hold an extractable array.

slint_interpreter_value_to_bool

slint_interpreter_value_to_brush

slint_interpreter_value_to_image

slint_interpreter_value_to_model

If the value contains a model set from [slint_interpreter_value_new_model] with the same vtable pointer, return the model that was set. Returns a null ptr otherwise

slint_interpreter_value_to_number

slint_interpreter_value_to_string

slint_interpreter_value_to_struct

slint_interpreter_value_type