slint_interpreter

Function register_font_from_memory

pub fn register_font_from_memory(
    font_data: impl AsRef<[u8]> + Send + Sync + 'static,
) -> Result<FontHandle, RegisterFontError>

Available on crate feature std only.

Register a font from byte data dynamically.

The argument is the owner of a slice of byte that contains the data of the font. It can for example be a &'static [u8] obtained from include_bytes!() or a Vec<u8> obtained from std::fs::read().

The data must be in the TrueType (TTF) format.

Example

let font_bytes = include_bytes!("sharedfontique/DejaVuSans.ttf");
slint::register_font_from_memory(font_bytes).expect("Failed to register font");