Struct slint_interpreter::ComponentDefinition

pub struct ComponentDefinition { /* private fields */ }

ComponentDefinition is a representation of a compiled component from .slint markup.

It can be constructed from a .slint file using the Compiler::build_from_path or Compiler::build_from_source functions. And then it can be instantiated with the Self::create function.

The ComponentDefinition acts as a factory to create new instances. When youโve finished creating the instances it is safe to drop the ComponentDefinition.

Implementations

impl ComponentDefinition

pub fn create(&self) -> Result<ComponentInstance, PlatformError>

Creates a new instance of the component and returns a shared handle to it.

pub fn properties(&self) -> impl Iterator<Item = (String, ValueType)> + '_

Returns an iterator over all publicly declared properties. Each iterator item is a tuple of property name and property type for each of them.

pub fn callbacks(&self) -> impl Iterator<Item = String> + '_

Returns the names of all publicly declared callbacks.

pub fn functions(&self) -> impl Iterator<Item = String> + '_

Returns the names of all publicly declared functions.

pub fn globals(&self) -> impl Iterator<Item = String> + '_

Returns the names of all exported global singletons

Note: Only globals that are exported or re-exported from the main .slint file will be exposed in the API

pub fn global_properties( &self, global_name: &str, ) -> Option<impl Iterator<Item = (String, ValueType)> + '_>

List of publicly declared properties in the exported global singleton specified by its name.

pub fn global_callbacks( &self, global_name: &str, ) -> Option<impl Iterator<Item = String> + '_>

List of publicly declared callbacks in the exported global singleton specified by its name.

pub fn global_functions( &self, global_name: &str, ) -> Option<impl Iterator<Item = String> + '_>

List of publicly declared functions in the exported global singleton specified by its name.

pub fn name(&self) -> &str

The name of this Component as written in the .slint file

pub fn type_loader(&self) -> Rc<TypeLoader>

Return the TypeLoader used when parsing the code in the interpreter.

WARNING: this is not part of the public API

pub fn raw_type_loader(&self) -> Option<TypeLoader>

Return the TypeLoader used when parsing the code in the interpreter in a state before most passes were applied by the compiler.

Each returned type loader is a deep copy of the entire state connected to it, so this is a fairly expensive function!

WARNING: this is not part of the public API