Enum slint_interpreter::Value

pub enum Value {
    Void,
    Number(f64),
    String(SharedString),
    Bool(bool),
    Image(Image),
    Model(ModelRc<Value>),
    Struct(Struct),
    Brush(Brush),
    // some variants omitted
}

This is a dynamically typed value used in the Slint interpreter. It can hold a value of different types, and you should use the From or TryFrom traits to access the value.

use core::convert::TryInto;
// create a value containing an integer
let v = Value::from(100u32);
assert_eq!(v.try_into(), Ok(100u32));

Variants

Void

There is nothing in this value. Thatโs the default. For example, a function that do not return a result would return a Value::Void

Number(f64)

An int or a float (this is also used for unit based type such as length or angle)

String(SharedString)

Correspond to the string type in .slint

Bool(bool)

Correspond to the bool type in .slint

Image(Image)

Correspond to the image type in .slint

Model(ModelRc<Value>)

A model (that includes array in .slint)

Struct(Struct)

An object

Brush(Brush)

Correspond to brush or color type in .slint. For color, this is then a Brush::SolidColor

Implementations

impl Value

pub fn value_type(&self) -> ValueType

Returns the type variant that this value holds without the containing value.